requirements.txt
edge-tts>=6.1.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pystray>=0.19.0
Pillow>=10.0.0
